# ApiGateway/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Agriculture Crop System API",
    "description": "To help farmers sell their produce via web"
  },
  "host": "localhost:4545",
  "basePath": "/",
...
  "paths": {
...
    "/farmer/signup": {
      "post": {
        "tags":[
          "Authentication API"
        ],
        "description": "Create new Farmer in the system",
        "parameters": [
          {
            "name": "Farmer",
            "in":"body",
            "description": "New Farmer that we want to create",
            "schema": {
              "$ref": "#/definitions/Signup"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "responses": {
          "200":{
            "description": "New Farmer Created"
           
          }
        }
      }
    },
    "/farmer/login":{
      "post":{
        "tags":[
          "Authentication API"
        ],
        "description": "Login with new Farmer in the system",
        "parameters": [
          {
            "name": "Farmer",
            "in":"body",
            "description": "Details for logging in the farmer",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "produces":[
          "application/json"
        ],
        "responses": {
          "200":{
            "description": "Farmer Logged in"
            
          }
        }
      }
    },
...
             "/farmerprofile/{id}": {
              "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "description": "ID of farmer to be updated",
                  "type": "string"
                }
              ],
              "put": {
                "summary": "Updates the information",
                "tags": [
                  "Farmer Service"
                ],
                "parameters": [
                  {"name": "Update Profile",
                    "in": "body",
                    "description": "Farmer with new data",
                    "schema": {
                      "$ref": "#definitions/Profile"
                    }
                  }

                ],
                "responses": {
                  "200": {
                    "description": "Profile updated successfully",
                    "schema": {
                      "$ref": "#/definitions/Profile"
                    }
                   
                  },
                  "404": {
                    "description": "Farmer not found!"
                  }
                }
              }
            },
...
  },
  "definitions": {
...
      "Signup":{
        "properties": {
          
          "name":{
            "type": "string"
          },
          "email":{
            "type": "string"
          },
          "password":{
            "type": "string"
          }
            
          }
        },
        "Login":{
          "properties": {
           
            "email":{
              "type": "string"
            },
            "password":{
              "type": "string"
            }
              
            }
          },
...
           "Profile":{
             "properties": {
               "Name":{
                 "type":"string"
               },
               "ContactNumber":{
                 "type":"integer"
               },
               "Address":{
                 "type":"object",
                 "items": {
                   
                   "properties": {
                     "Street":{
                       "type":"string"
                     },
                     "City":{
                      "type":"string"
                    },
                    "State":{
                      "type":"string"
                    },
                    "ZipCode":{
                      "type":"integer"
                    }
                   }
                 }
               },
               "Bankdetails":{
                 "type":"object",
                 "items": {
                   "properties": {
                    "accNum":{
                      "type":"integer"
                    },
                    "ifscCode":{
                      "type":"string"
                    } 
                   }
                 }
               }
             }
           } 
    }
}